+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 1.1.1
 - Added extension toNullable to Option
 - Added errorMessage and internalErrorCode optional parameters to trayCatch
+
+## 1.2.1
+- Deprecated Validation. Use validatedResult instead  
+- Added parallelism to Iterable and ValidatedResult<Iterable>
+- Added try_ to validate a function that can throws Errors or Exceptions